+++ lib/src/commonMain/kotlin/de/jhoopmann/topmostwindow/compose/ui/awt/ComposeTopMostWindow.kt
@@ -3,13 +3,15 @@ package de.jhoopmann.topmostwindow.compose.ui.awt
 import androidx.compose.ui.awt.ComposeWindow
 import de.jhoopmann.topmostwindow.awt.ui.TopMost
 import de.jhoopmann.topmostwindow.awt.ui.TopMostImpl
+import java.awt.Window
 
-open class ComposeTopMostWindow(
-    open val update: (ComposeWindow) -> Unit,
-    open val composeWindow: ComposeWindow = ComposeWindow(),
-    open val delegate: TopMostImpl = TopMostImpl()
+class ComposeTopMostWindow internal constructor(
+    val composeWindow: ComposeWindow = ComposeWindow(),
+    private val delegate: TopMostImpl = TopMostImpl()
 ) : TopMost by delegate {
-    open fun setVisible(visible: Boolean) {
+    internal lateinit var update: (ComposeWindow) -> Unit
+
+    fun setVisible(visible: Boolean) {
         if (visible) {
             update.invoke(composeWindow)
         }