+++ lib/src/commonMain/kotlin/de/jhoopmann/topmostwindow/compose/ui/window/TopMostWindow.types.kt
@@ -1,0 +1,5 @@
+package de.jhoopmann.topmostwindow.compose.ui.window
+
+import de.jhoopmann.topmostwindow.compose.ui.awt.ComposeTopMostWindow
+
+typealias OnCreateTopMostWindowEvent = ((ComposeTopMostWindow) -> Unit)?